+++ .eslintrc.cjs
@@ -35,9 +35,9 @@ module.exports = {
       rules: {},
     },
     {
-      files: ["*.test.js"],
+      files: ['*.test.js'],
       rules: {
-        "no-unused-expressions": "off"
+        'no-unused-expressions': 'off',
       }
     }
   ],