+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@adobe/spacecat-shared-google-client-v1.1.16](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.1.15...@adobe/spacecat-shared-google-client-v1.1.16) (2024-07-27)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#304](https://github.com/adobe/spacecat-shared/issues/304)) ([c6c56a7](https://github.com/adobe/spacecat-shared/commit/c6c56a72897acb60fb042215b708816ec16a5870))
+
+# [@adobe/spacecat-shared-google-client-v1.1.15](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.1.14...@adobe/spacecat-shared-google-client-v1.1.15) (2024-07-22)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#295](https://github.com/adobe/spacecat-shared/issues/295)) ([4e8d50d](https://github.com/adobe/spacecat-shared/commit/4e8d50d8d88b2d68b483fda2ead31542d15ee952))
+
 # [@adobe/spacecat-shared-google-client-v1.1.14](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.1.13...@adobe/spacecat-shared-google-client-v1.1.14) (2024-07-13)